requirements.txt
fastapi
uvicorn
streamlit
pandas
numpy
google-generativeai
python-dotenv
plotly
sqlalchemy
